Use prefix and suffix slots for icons
Fix #37

As a consequence, the order of child nodes no more respects the
iconAfterText property (it does not need to).

Also, text content is no more wrapped inside a <slot> when using both
icon and text content.

It's not clear whether this is a breaking change or a bug fix, as users
might have applied styling which depends on the old behavior.